vehicle-parking-app/backend/requirements.txt
# Flask and core dependencies
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-CORS==4.0.0
Werkzeug==2.3.7

# Database
SQLAlchemy==2.0.21

# Authentication and security
PyJWT==2.8.0
bcrypt==4.0.1

# Redis and Caching
redis==5.0.0
Flask-Caching==2.1.0

# Celery for background tasks
celery==5.3.2
celery[redis]==5.3.2

# Email functionality
secure-smtplib==0.1.1

# HTTP requests (for webhooks)
requests==2.31.0

# SMS functionality (Twilio)
twilio==8.5.0

# Environment variables
python-dotenv==1.0.0

# Date and time utilities
python-dateutil==2.8.2

# Data processing
pandas==2.1.1
numpy==1.25.2

# Development dependencies
pytest==7.4.2
pytest-flask==1.2.0
python-decouple==3.8

# Production server
gunicorn==21.2.0

# File handling
openpyxl==3.1.2

# Validation
marshmallow==3.20.1
marshmallow-sqlalchemy==0.29.0

# Monitoring and logging
sentry-sdk[flask]==1.32.0

# API documentation (optional)
flask-restx==1.2.0

# Rate limiting
Flask-Limiter==3.5.0
